Increase `adb devices -l` max output size
For simplicity, the parsing of `adb devices -l` output is done in a
single pass on the whole output.

This output was limited to 4096 bytes. Since there are about 100 chars
per device line, this limited the number of connected devices to ~40.

Increase to 65536 bytes to avoid a limitation in practice.